Update swc.css
In the html view of the lessons, the table headers are very difficult to read because the default header color of dark blue is very close to the default font color. I've simply added a rule that lightens the blue when the header style appears inside a table row. I'm not sure if this is the best way to fix the problem - so I won't be offended if this change is rejected :)