07-join.md: Don't explain distinction between WHERE and ON
The distinction only matters for outer joins [1], and we never discuss
outer joins.  Here's an example of the distinction for outer joins:

  sqlite> SELECT Site.name, Visited.site, Visited.ident, Visited.dated
     ...> FROM Site
     ...> LEFT OUTER JOIN Visited ON Site.name=Visited.site AND Visited.dated IS NULL;
  DR-1|||
  DR-3|DR-3|752|
  MSK-4|||
  sqlite> SELECT Site.name, Visited.site, Visited.ident, Visited.dated
     ...> FROM Site
     ...> LEFT JOIN Visited ON Site.name=Visited.site
     ...> WHERE Visited.dated IS NULL;
  DR-3|DR-3|752|

So the difference (where it matters for outer joins) is that a left
outer join (for example) will return all rows from the left table even
if the ON clause never matches [2], while all returned rows must match
the WHERE clause.  This is a semantic difference in the *meaning* of
the filtering and has nothing to do with when the filtering happens.

With this commit we acknowledge that there is a difference, but
explain that it doesn't matter for this lesson, and give them a
keyword and links to further docs so they hear that it may matter to
them eventually and have some idea of when they'll need to do further
research.  I don't think it's worth trying to explain the difference
until you can show an outer join like the above example to illustrate
it.

For outer-join docs, I've linked to Wikipedia, because the description
there:

  An outer join does not require each record in the two joined tables
  to have a matching record...

seems more accessible than the SQLite docs [3]:

  If the join-operator is a "LEFT JOIN" or "LEFT OUTER JOIN", then
  after the ON or USING filtering clauses have been applied, an extra
  row is added to the output for each row in the original left-hand
  input dataset that corresponds to no rows at all in the composite
  dataset (if any). The added rows contain NULL values in the columns
  that would normally contain values copied from the right-hand input
  dataset.

[1]: http://stackoverflow.com/a/354094
[2]: http://en.wikipedia.org/wiki/Join_%28SQL%29#Left_outer_join
[3]: https://www.sqlite.org/lang_select.html#fromclause