data/visited.csv: Fix '1939' -> '1930' typo
With:

  $ sed -i 's/1939/1930/g' *.md code/gen-survey-database.sql \
  >   data/visited.csv fig/*.svg

In index.md, we have:

  In the late 1920s and early 1930s, ...

So 1939 is probably a fat-finger error dating back to 61ae6ffb (First
three lessons on SQL for beginners, 2013-11-09).  It's possible that
the error was intentional, to setup an excuse to talk about dirty
data, or OCR errors, or something like that.  We already talk about
potentially dirty data with units and the anomalous salinity
measurements in 08-hygiene.md though, so I don't think the 1939 date
is worth the trouble.