Clarified ORDER BY
While prepping for the teaching lesson tomorrow, I saw:

"By default, when we use ORDER BY
results are sorted in ascending order " Which is a very worrying statement. While we know that that "default" refers to the ORDER BY clause, I suspect that this will trip up new readers by contradicting the paragraph above of "nothing's returned in any particular order." I added a bit of clarification and an extra warning paragraph. That paragraph might be put into some sort of aside-box, but I'm not sure which one to use.