Fix the South Pole filtering question
Hi all,

(I'm submitting this as part of my check procedure)

We noticed this problem during a recent workshop. The question is worded incorrectly because it is assuming the poles are at 0° when in fact they are at 90°. 

You could actually resolve this question in a number of ways depending on what you want to ask. The change I am submitting does the least amount of altering to the SQL and the ensuing results less compared to other possible solutions and still maintains the spirit of the question, i.e. it filters some of them and leaves others by changing from an `OR` to an `AND`.